mover contacto a su propia vista
